# ui/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Scanopy API",
    "description": "Network topology discovery and visualization API",
    "license": {
      "name": "Dual (AGPL3.0, Commercial License Available)"
    },
    "version": "0.12.5"
  },
  "paths": {
...
    "/api/invites": {
      "get": {
        "tags": [
          "invites"
        ],
        "summary": "List all invites for organization",
        "operationId": "get_invites",
        "responses": {
          "200": {
            "description": "List of active invites",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Invite"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "post": {
        "tags": [
          "invites"
        ],
        "summary": "Create an organization invite",
        "operationId": "create_invite",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInviteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Invite created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invite"
                }
              }
            }
          },
          "400": {
            "description": "User already has an account"
          },
          "403": {
            "description": "Seat limit reached or insufficient permissions"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/invites/{id}": {
      "get": {
        "tags": [
          "invites"
        ],
        "summary": "Get an invite by ID",
        "operationId": "get_invite",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Invite ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invite details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invite"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired invite"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/invites/{id}/revoke": {
      "delete": {
        "tags": [
          "invites"
        ],
        "summary": "Revoke an invite",
        "operationId": "revoke_invite",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Invite ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invite revoked"
          },
          "400": {
            "description": "Invalid invite"
          },
          "403": {
            "description": "Cannot revoke this invite"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateInviteRequest": {
        "type": "object",
        "required": [
          "permissions",
          "network_ids"
        ],
        "properties": {
          "expiration_hours": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "network_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "permissions": {
            "$ref": "#/components/schemas/UserOrgPermissions"
          },
          "send_to": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "Invite": {
        "allOf": [
          {
            "$ref": "#/components/schemas/InviteBase"
          },
          {
            "type": "object",
            "properties": {
              "created_at": {
                "type": "string",
                "format": "date-time",
                "readOnly": true
              },
              "id": {
                "type": "string",
                "format": "uuid",
                "readOnly": true
              },
              "updated_at": {
                "type": "string",
                "format": "date-time",
                "readOnly": true
              }
            }
          }
        ]
      },
      "InviteBase": {
        "type": "object",
        "required": [
          "organization_id",
          "permissions",
          "network_ids",
          "url",
          "created_by",
          "expires_at"
        ],
        "properties": {
          "created_by": {
            "type": "string",
            "format": "uuid"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "network_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "organization_id": {
            "type": "string",
            "format": "uuid"
          },
          "permissions": {
            "$ref": "#/components/schemas/UserOrgPermissions"
          },
          "send_to": {
            "type": [
              "string",
              "null"
            ]
          },
          "url": {
            "type": "string"
          }
        }
      },
...
      "UserOrgPermissions": {
        "type": "string",
        "enum": [
          "Owner",
          "Admin",
          "Member",
          "Viewer"
        ]
      }
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for daemon authentication. Generate keys in the web UI under Settings > API Keys."
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    }
  }
}